move files to different directories
