+++ app/app/src/main/java/com/ar/enbaldeapp/services/reply/IServerReply.java
@@ -1,4 +1,7 @@
-package com.ar.enbaldeapp.services;
+package com.ar.enbaldeapp.services.reply;
+
+import com.ar.enbaldeapp.services.ApiError;
+import com.ar.enbaldeapp.services.ApiResponse;
 
 import java.io.IOException;
