+++ app/app/src/main/java/com/ar/enbaldeapp/services/reply/ServerReply.java
@@ -1,4 +1,6 @@
-package com.ar.enbaldeapp.services;
+package com.ar.enbaldeapp.services.reply;
+
+import com.ar.enbaldeapp.services.reply.IServerReply;
 
 import java.io.BufferedReader;
 import java.io.IOException;
