+++ app/app/src/main/java/com/ar/enbaldeapp/services/requesters/IRequester.java
@@ -1,7 +1,7 @@
 package com.ar.enbaldeapp.services.requesters;
 
-import com.ar.enbaldeapp.services.IHttpUrlConnectionWrapper;
-import com.ar.enbaldeapp.services.IServerReply;
+import com.ar.enbaldeapp.services.connection.IHttpUrlConnectionWrapper;
+import com.ar.enbaldeapp.services.reply.IServerReply;
 
 import java.io.IOException;
